requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
easyocr>=1.6.0
sympy>=1.12
python-dotenv>=1.0.0
google-generativeai>=0.3.0
